video-consultation-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.35
asyncpg==0.30.0
pydantic==2.8.0
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-socketio==5.11.0
aio-pika==9.3.1
alembic==1.13.1
python-dotenv==1.0.1
httpx==0.26.0
psycopg2-binary==2.9.10